backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "moduga-jakga"
version = "0.1.0"
description = "모두가 작가 - 시각장애인을 위한 AI 기반 글쓰기 플랫폼 백엔드"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "StartFolio Team" },
]

dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "uvicorn[standard]>=0.30.0",
    "supabase>=2.0.0",
    "openai>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "python-docx>=1.1.0",
    "ebooklib>=0.18",
    "Pillow>=10.0.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
